requirements.txt
# Core dependencies for Neuralux AI Layer

# Message Bus
nats-py>=2.7.0
asyncio-nats-client>=0.11.5

# AI/ML Frameworks
torch>=2.1.0
transformers>=4.36.0
# NOTE: llama-cpp-python will be installed with GPU support if CUDA is detected during install.sh
# For manual GPU setup: CMAKE_ARGS="-DGGML_CUDA=on" pip install llama-cpp-python --upgrade --force-reinstall --no-cache-dir
llama-cpp-python>=0.2.0
sentence-transformers>=2.2.0

# Vector Store & Databases
qdrant-client>=1.7.0
duckdb>=0.9.0
sqlite-vss>=0.1.0

# API & Web
fastapi>=0.109.0
uvicorn>=0.27.0
aiohttp>=3.9.0
python-multipart>=0.0.6  # For FastAPI file uploads
grpcio>=1.60.0
grpcio-tools>=1.60.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# System Integration
psutil>=5.9.0
pyinotify>=0.9.6
watchdog>=2.3.0  # For filesystem event monitoring
dbus-python>=1.3.2
pynvml>=11.5.0

# CLI & UI
rich>=13.7.0
click>=8.1.7
prompt-toolkit>=3.0.43
typer>=0.9.0
rapidfuzz>=3.5.2
python-xlib>=0.33
duckduckgo-search>=5.3.0
ddgs>=1.0.6
beautifulsoup4>=4.12.3

# Audio Processing
openai-whisper>=20231117
faster-whisper>=0.10.0
piper-tts>=1.2.0
onnxruntime>=1.16.0
soundfile>=0.12.0
pyaudio>=0.2.13
torchaudio>=2.0.0
nvidia-cudnn-cu12>=9.0.0  # For CUDA-accelerated STT

# Document Parsing (for filesystem service)
pypdf>=3.17.0
python-docx>=1.1.0
odfpy>=1.4.1
openpyxl>=3.1.2

# Utilities
python-dotenv>=1.0.0
pyyaml>=6.0.1
aiofiles>=23.2.1
httpx>=0.26.0
tenacity>=8.2.3
structlog>=24.1.0
redis>=5.0.0

# Vision / OCR
Pillow>=10.0.0
opencv-python-headless>=4.10.0.84
pytesseract>=0.3.10
PaddlePaddle>=2.6.1; platform_system == "Linux"
paddleocr>=2.7.0
mss>=9.0.1

# Development
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.12.0
ruff>=0.1.0
mypy>=1.8.0

# Image Generation (for Vision Service)
diffusers>=0.24.0
accelerate>=0.25.0
safetensors>=0.4.0
sentencepiece>=0.1.99
protobuf>=3.20.0
bitsandbytes>=0.41.0  # For 8-bit quantization (saves VRAM)
